[frontend] Rename
