+++ openbas-front/src/actions/mapper/mapper.ts
